+++ offchain/crates/solana-cli/src/command/shreds/validator_client_rewards/claim.rs
@@ -86,8 +86,7 @@ impl ClaimCommand {
             .connection_options
             .clone()
             .into_shred_subscription_connection();
-        let mut wallet = Wallet::try_from(self.solana_payer_options)?;
-        wallet.connection = dz_connection;
+        let wallet = Wallet::try_new(self.solana_payer_options, Some(dz_connection))?;
         let wallet_key = wallet.pubkey();
 
         // Derive every PDA we need.