+++ offchain/crates/solana-cli/src/command/shreds/validator_client_rewards/show.rs
@@ -136,8 +136,7 @@ impl ShowCommand {
             .connection_options
             .clone()
             .into_shred_subscription_connection();
-        let mut wallet = Wallet::try_from(self.solana_payer_options)?;
-        wallet.connection = dz_connection;
+        let wallet = Wallet::try_new(self.solana_payer_options, Some(dz_connection))?;
 
         let vcr_key = find_validator_client_rewards_address(self.client_id).0;
         let vcr_account = wallet